Refactor logsScreen hook to own file
